move platform.h from device to top-level
