+++ src/TiledArray/device/cpu_cuda_vector.h
@@ -4,8 +4,8 @@
 
 #include <btas/array_adaptor.h>
 
-#include <TiledArray/device/platform.h>
 #include <TiledArray/device/thrust.h>
+#include <TiledArray/platform.h>
 
 #include <madness/world/archive.h>
 
@@ -213,9 +213,9 @@ struct ArchiveLoadImpl<Archive, TiledArray::cpu_cuda_vector<T>> {
   static inline void load(const Archive& ar,
                           TiledArray::cpu_cuda_vector<T>& x) {
     typename TiledArray::cpu_cuda_vector<T>::size_type n(0);
-    ar& n;
+    ar & n;
     x.resize(n);
-    for (auto& xi : x) ar& xi;
+    for (auto& xi : x) ar & xi;
   }
 };
 
@@ -223,8 +223,8 @@ template <class Archive, typename T>
 struct ArchiveStoreImpl<Archive, TiledArray::cpu_cuda_vector<T>> {
   static inline void store(const Archive& ar,
                            const TiledArray::cpu_cuda_vector<T>& x) {
-    ar& x.size();
-    for (const auto& xi : x) ar& xi;
+    ar & x.size();
+    for (const auto& xi : x) ar & xi;
   }
 };
